+++ engine/assets/ObjLoader.cpp
@@ -16,25 +16,25 @@ namespace ouzel
     {
         namespace
         {
-            constexpr auto isWhitespace(std::uint8_t c)
+            constexpr auto isWhitespace(std::byte c)
             {
                 return static_cast<char>(c) == ' ' ||
                     static_cast<char>(c) == '\t';
             }
 
-            constexpr auto isNewline(std::uint8_t c)
+            constexpr auto isNewline(std::byte c)
             {
                 return static_cast<char>(c) == '\r' ||
                     static_cast<char>(c) == '\n';
             }
 
-            constexpr auto isControlChar(std::uint8_t c)
+            constexpr auto isControlChar(std::byte c)
             {
-                return c <= 0x1F;
+                return static_cast<std::uint8_t>(c) <= 0x1F;
             }
 
-            void skipWhitespaces(std::vector<std::uint8_t>::const_iterator& iterator,
-                                 std::vector<std::uint8_t>::const_iterator end)
+            void skipWhitespaces(std::vector<std::byte>::const_iterator& iterator,
+                                 std::vector<std::byte>::const_iterator end)
             {
                 while (iterator != end)
                     if (isWhitespace(*iterator))
@@ -43,8 +43,8 @@ namespace ouzel
                         break;
             }
 
-            void skipLine(std::vector<std::uint8_t>::const_iterator& iterator,
-                          std::vector<std::uint8_t>::const_iterator end)
+            void skipLine(std::vector<std::byte>::const_iterator& iterator,
+                          std::vector<std::byte>::const_iterator end)
             {
                 while (iterator != end)
                 {
@@ -58,8 +58,8 @@ namespace ouzel
                 }
             }
 
-            std::string parseString(std::vector<std::uint8_t>::const_iterator& iterator,
-                                    std::vector<std::uint8_t>::const_iterator end)
+            std::string parseString(std::vector<std::byte>::const_iterator& iterator,
+                                    std::vector<std::byte>::const_iterator end)
             {
                 std::string result;
 
@@ -76,8 +76,8 @@ namespace ouzel
                 return result;
             }
 
-            std::int32_t parseInt32(std::vector<std::uint8_t>::const_iterator& iterator,
-                                    std::vector<std::uint8_t>::const_iterator end)
+            std::int32_t parseInt32(std::vector<std::byte>::const_iterator& iterator,
+                                    std::vector<std::byte>::const_iterator end)
             {
                 std::string value;
                 std::uint32_t length = 1;
@@ -104,8 +104,8 @@ namespace ouzel
                 return std::stoi(value);
             }
 
-            float parseFloat(std::vector<std::uint8_t>::const_iterator& iterator,
-                             std::vector<std::uint8_t>::const_iterator end)
+            float parseFloat(std::vector<std::byte>::const_iterator& iterator,
+                             std::vector<std::byte>::const_iterator end)
             {
                 std::string value;
                 std::uint32_t length = 1;
@@ -176,8 +176,8 @@ namespace ouzel
                 return std::stof(value);
             }
 
-            bool parseToken(const std::vector<std::uint8_t>& str,
-                            std::vector<std::uint8_t>::const_iterator& iterator,
+            bool parseToken(const std::vector<std::byte>& str,
+                            std::vector<std::byte>::const_iterator& iterator,
                             char token)
             {
                 if (iterator == str.end() || static_cast<char>(*iterator) != token) return false;
@@ -195,7 +195,7 @@ namespace ouzel
 
         bool ObjLoader::loadAsset(Bundle& bundle,
                                   const std::string& name,
-                                  const std::vector<std::uint8_t>& data,
+                                  const std::vector<std::byte>& data,
                                   bool mipmaps)
         {
             std::string objectName = name;
